* Added services directory for functions that can be reused across different routes: bitbox, requestUtils * Restructured block.ts with the use of services from previous point * Fix of typescript errors (added new lib to tsconfig.json)
