+++ src/routes/v2/services/bitbox.ts
@@ -1,0 +1,7 @@
+import axios from "axios"
+
+const BitboxHTTP = axios.create({
+  baseURL: process.env.RPC_BASEURL
+})
+
+export const getInstance = () => BitboxHTTP
