+++ tsconfig.json
@@ -6,7 +6,7 @@
     "module": "commonjs",
     "target": "es5",
     "allowJs": true,
-    "lib": ["es2015"],
+    "lib": ["es2015", "es2015.symbol", "dom"],
     "typeRoots": ["node_modules/@types/"]
   },
   "include": ["src/**/*"],
